@@ -1,15 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 19
-    buildToolsVersion '20.0.0'
+    compileSdkVersion 21
+    buildToolsVersion '21.1.1'
 
     defaultConfig {
-        applicationId "yuku.alkitabfeedback"
         minSdkVersion 14
-        targetSdkVersion 19
-        versionCode 1
-        versionName "1.0"
+        targetSdkVersion 21
     }
 
     compileOptions {
@@ -26,6 +23,6 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:support-v4:20.0.0'
-    compile 'com.squareup.okhttp:okhttp:2.0.0'
+    compile 'com.android.support:support-v4:21.0.2'
+    compile 'com.squareup.okhttp:okhttp:2.1.0'
 }
