@@ -1,13 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 19
-    buildToolsVersion "20.0.0"
+    compileSdkVersion 21
+    buildToolsVersion '21.1.1'
 
     defaultConfig {
-        applicationId "yuku.alkitabconverter.util"
         minSdkVersion 14
-        targetSdkVersion 19
+        targetSdkVersion 21
     }
     buildTypes {
         release {
