@@ -1,6 +1,5 @@
 package yuku.alkitab.base.ac;
 
-import android.app.ActionBar;
 import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.Context;
@@ -9,20 +8,19 @@ import android.database.Cursor;
 import android.database.MatrixCursor;
 import android.os.AsyncTask;
 import android.os.Bundle;
+import android.support.v4.widget.CursorAdapter;
+import android.support.v7.widget.SearchView;
+import android.support.v7.widget.Toolbar;
 import android.text.SpannableStringBuilder;
 import android.text.TextUtils;
 import android.util.SparseBooleanArray;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.inputmethod.InputMethodManager;
-import android.widget.AdapterView;
 import android.widget.AutoCompleteTextView;
-import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.CompoundButton;
-import android.widget.CursorAdapter;
 import android.widget.ListView;
-import android.widget.SearchView;
 import android.widget.TextView;
 import android.widget.Toast;
 import yuku.afw.V;
@@ -58,7 +56,7 @@ public class SearchActivity extends BaseActivity {
 
 	final String COLUMN_QUERY_STRING = "query_string";
 
-	Button bVersion;
+	TextView bVersion;
 	SearchView searchView;
 	ListView lsSearchResults;
 	View empty;
@@ -133,12 +131,7 @@ public class SearchActivity extends BaseActivity {
 			}
 
 			// sometimes this is called from bg. So we need to make sure this is run on UI thread.
-			runOnUiThread(new Runnable() {
-				@Override
-				public void run() {
-					swapCursor(mc);
-				}
-			});
+			runOnUiThread(() -> swapCursor(mc));
 		}
 	}
 
@@ -163,12 +156,12 @@ public class SearchActivity extends BaseActivity {
 		tFilterAdvanced = V.get(this, R.id.tFilterAdvanced);
 		bEditFilter = V.get(this, R.id.bEditFilter);
 
-		final ActionBar actionBar = getActionBar();
-		actionBar.setDisplayShowCustomEnabled(true);
+		final Toolbar toolbar = V.get(this, R.id.toolbar);
+		setSupportActionBar(toolbar); // must be done first before below lines
+		toolbar.setNavigationIcon(R.drawable.abc_ic_ab_back_mtrl_am_alpha);
+		toolbar.setNavigationOnClickListener(v -> navigateUp());
 
-		final View actionCustomView = getLayoutInflater().cloneInContext(actionBar.getThemedContext()).inflate(R.layout.activity_search_action_custom_view, null);
-		bVersion = V.get(actionCustomView, R.id.bVersion);
-		actionBar.setCustomView(actionCustomView);
+		bVersion = V.get(this, R.id.bVersion);
 
 		searchInVersion = S.activeVersion;
 		searchInVersionId = S.activeVersionId;
@@ -232,18 +225,11 @@ public class SearchActivity extends BaseActivity {
 		
 		hiliteColor = U.getHighlightColorByBrightness(S.applied.backgroundBrightness);
 
-		lsSearchResults.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-			@Override
-			public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-				int ari = adapter.getSearchResults().get(position);
-				startActivity(Launcher.openAppAtBibleLocationWithVerseSelected(ari));
-			}
+		lsSearchResults.setOnItemClickListener((parent, view, position, id) -> {
+			int ari = adapter.getSearchResults().get(position);
+			startActivity(Launcher.openAppAtBibleLocationWithVerseSelected(ari));
 		});
-		bEditFilter.setOnClickListener(new View.OnClickListener() {
-			@Override public void onClick(View v) {
-				bEditFilter_click();
-			}
-		});
+		bEditFilter.setOnClickListener(v -> bEditFilter_click());
 		cFilterOlds.setOnCheckedChangeListener(cFilterOlds_checkedChange);
 		cFilterNews.setOnCheckedChangeListener(cFilterNews_checkedChange);
 		cFilterSingleBook.setOnCheckedChangeListener(cFilterSingleBook_checkedChange);
@@ -581,14 +567,11 @@ public class SearchActivity extends BaseActivity {
 						sb.append(fallback);
 
 						tSearchTips.setText(sb);
-						tSearchTips.setOnClickListener(new View.OnClickListener() {
-							@Override
-							public void onClick(final View v) {
-								if (Ari.toVerse(fallbackAri) == 0) {
-									startActivity(Launcher.openAppAtBibleLocation(fallbackAri));
-								} else {
-									startActivity(Launcher.openAppAtBibleLocationWithVerseSelected(fallbackAri));
-								}
+						tSearchTips.setOnClickListener(v -> {
+							if (Ari.toVerse(fallbackAri) == 0) {
+								startActivity(Launcher.openAppAtBibleLocation(fallbackAri));
+							} else {
+								startActivity(Launcher.openAppAtBibleLocationWithVerseSelected(fallbackAri));
 							}
 						});
 					} else {
