@@ -4,27 +4,24 @@ import android.app.AlertDialog;
 import android.app.LoaderManager;
 import android.app.ProgressDialog;
 import android.content.AsyncTaskLoader;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.Loader;
 import android.os.Bundle;
 import android.os.Parcel;
 import android.os.Parcelable;
+import android.support.v7.widget.SearchView;
+import android.support.v7.widget.Toolbar;
 import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.View.OnClickListener;
 import android.view.ViewGroup;
 import android.view.Window;
 import android.widget.AdapterView;
 import android.widget.BaseAdapter;
-import android.widget.Button;
 import android.widget.CheckBox;
-import android.widget.CompoundButton;
 import android.widget.ListView;
 import android.widget.PopupMenu;
-import android.widget.SearchView;
 import android.widget.TextView;
 import yuku.afw.App;
 import yuku.afw.V;
@@ -53,7 +50,7 @@ public class SongListActivity extends BaseActivity {
 	
 	SearchView searchView;
 	ListView lsSong;
-	Button bChangeBook;
+	TextView bChangeBook;
 	CheckBox cDeepSearch;
 	View panelFilter;
 	
@@ -132,14 +129,17 @@ public class SongListActivity extends BaseActivity {
 	}
 	
 	@Override protected void onCreate(Bundle savedInstanceState) {
-		requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
+		supportRequestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
 		super.onCreate(savedInstanceState);
-		setProgressBarIndeterminate(true);
-		
 		setContentView(R.layout.activity_song_list);
-		
-		setTitle(R.string.sn_songs_activity_title);
-		
+
+		final Toolbar toolbar = V.get(this, R.id.toolbar);
+		setSupportActionBar(toolbar); // must be done first before below lines
+		toolbar.setNavigationIcon(R.drawable.abc_ic_ab_back_mtrl_am_alpha);
+		toolbar.setNavigationOnClickListener(v -> navigateUp());
+
+		setSupportProgressBarIndeterminate(true);
+
 		searchView = V.get(this, R.id.searchView);
 		lsSong = V.get(this, R.id.lsSong);
 		bChangeBook = V.get(this, R.id.bChangeBook);
@@ -155,11 +155,14 @@ public class SongListActivity extends BaseActivity {
 		popupChangeBook = SongBookUtil.getSongBookPopupMenu(this, true, searchView);
 		popupChangeBook.setOnMenuItemClickListener(SongBookUtil.getSongBookOnMenuItemClickListener(songBookSelected));
 		
-		bChangeBook.setOnClickListener(bChangeBook_click);
-		cDeepSearch.setOnCheckedChangeListener(cDeepSearch_checkedChange);
+		bChangeBook.setOnClickListener(v -> popupChangeBook.show());
+		cDeepSearch.setOnCheckedChangeListener((buttonView, isChecked) -> startSearch());
 		
 		loader = new SongLoader();
-        
+
+		// initial
+		bChangeBook.setText(R.string.sn_bookselector_all);
+
         SearchState searchState = getIntent().getParcelableExtra(EXTRA_searchState);
         if (searchState != null) {
         	stillUsingInitialSearchState = true; { // prevent triggering
@@ -174,7 +177,7 @@ public class SongListActivity extends BaseActivity {
 	    			bChangeBook.setText(searchState.bookName);
 	    		}
         	} stillUsingInitialSearchState = false;
-        	setProgressBarIndeterminateVisibility(false); // somehow this is needed.
+        	setSupportProgressBarIndeterminateVisibility(false); // somehow this is needed.
         } else {
         	startSearch();
         }
@@ -186,12 +189,12 @@ public class SongListActivity extends BaseActivity {
         	
         	@Override public void onLoadFinished(Loader<List<SongInfo>> loader, List<SongInfo> data) {
         		adapter.setData(data);
-        		setProgressBarIndeterminateVisibility(false);
+        		setSupportProgressBarIndeterminateVisibility(false);
         	}
         	
         	@Override public void onLoaderReset(Loader<List<SongInfo>> loader) {
         		adapter.setData(null);
-        		setProgressBarIndeterminateVisibility(false);
+        		setSupportProgressBarIndeterminateVisibility(false);
         	}
         });
 	}
@@ -205,11 +208,7 @@ public class SongListActivity extends BaseActivity {
 		if (item.getItemId() == R.id.menuDeleteAll) {
 			new AlertDialog.Builder(this)
 			.setMessage(R.string.sn_delete_all_songs_explanation)
-			.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-				@Override public void onClick(DialogInterface dialog, int which) {
-					deleteAllSongs();
-				}
-			})
+			.setPositiveButton(R.string.ok, (dialog, which) -> deleteAllSongs())
 			.setNegativeButton(R.string.cancel, null)
 			.show();
 		}
@@ -224,17 +223,15 @@ public class SongListActivity extends BaseActivity {
 			@Override public void run() {
 				final int count = S.getSongDb().deleteAllSongs();
 				
-				runOnUiThread(new Runnable() {
-					@Override public void run() {
-						pd.dismiss();
-						
-						startSearch();
-						
-						new AlertDialog.Builder(SongListActivity.this)
-						.setMessage(getString(R.string.sn_delete_all_songs_result, count))
-						.setPositiveButton(R.string.ok, null)
-						.show();
-					}
+				runOnUiThread(() -> {
+					pd.dismiss();
+
+					startSearch();
+
+					new AlertDialog.Builder(SongListActivity.this)
+					.setMessage(getString(R.string.sn_delete_all_songs_result, count))
+					.setPositiveButton(R.string.ok, null)
+					.show();
 				});
 			}
 		}.start();
@@ -242,7 +239,7 @@ public class SongListActivity extends BaseActivity {
 
 	void startSearch() {
 		if (stillUsingInitialSearchState) return;
-		setProgressBarIndeterminateVisibility(true);
+		setSupportProgressBarIndeterminateVisibility(true);
 		loader.setFilterString(searchView.getQuery().toString());
 		loader.setDeepSearch(cDeepSearch.isChecked());
 		loader.forceLoad();
@@ -252,18 +249,6 @@ public class SongListActivity extends BaseActivity {
 		loader.setSelectedBookName(selectedBookName);
 		startSearch();
 	}
-	
-	OnClickListener bChangeBook_click = new OnClickListener() {
-		@Override public void onClick(View v) {
-			popupChangeBook.show();
-		}
-	};
-	
-	CompoundButton.OnCheckedChangeListener cDeepSearch_checkedChange = new CompoundButton.OnCheckedChangeListener() {
-		@Override public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-			startSearch();
-		}
-	};
 
 	SongBookUtil.OnSongBookSelectedListener songBookSelected = new SongBookUtil.OnSongBookSelectedListener() {
 		@Override public void onSongBookSelected(boolean all, SongBookUtil.SongBookInfo songBookInfo) {
