@@ -73,7 +73,7 @@ public class FontManagerActivity extends BaseActivity implements DownloadListene
 	};
 
 	@Override protected void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
+		super.onCreateWithNonToolbarUpButton(savedInstanceState);
 
 		setContentView(R.layout.activity_font_manager);
 		setTitle(R.string.fm_activity_title);
