@@ -55,7 +55,7 @@ public class MarkersActivity extends BaseActivity {
 	}
 
 	@Override protected void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
+		super.onCreateWithNonToolbarUpButton(savedInstanceState);
 		
 		setContentView(R.layout.activity_markers);
 		setTitle(R.string.activity_title_markers);
