@@ -1,55 +1,61 @@
 package yuku.alkitab.base.ac.base;
 
-import android.app.ActionBar;
+import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import android.support.v4.app.NavUtils;
+import android.support.v4.app.TaskStackBuilder;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.ActionBarActivity;
 import android.view.MenuItem;
-import com.google.analytics.tracking.android.EasyTracker;
+import yuku.alkitab.debug.R;
 
-public abstract class BaseActivity extends FragmentActivity {
+public abstract class BaseActivity extends ActionBarActivity {
 	public static final String TAG = BaseActivity.class.getSimpleName();
-	private boolean withUpButton;
 
-	/**
-	 * By default we put the up button on each activity.
-	 * And the behavior is to just finish() the current activity.
-	 * If we don't want this, e.g. on the root activity, change
-	 * super.onCreate(Bundle) to super.onCreate(Bundle, boolean)
-	 * on subclasses
-	 */
-	@Override protected void onCreate(Bundle savedInstanceState) {
-		onCreate(savedInstanceState, true);
-	}
-	
-	protected void onCreate(Bundle savedInstanceState, boolean withUpButton) {
-		this.withUpButton = withUpButton;
+	private boolean withNonToolbarUpButton;
+
+	@Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (Build.VERSION.SDK_INT >= 21) {
+            getWindow().setStatusBarColor(getResources().getColor(R.color.primary_dark));
+        }
+    }
+
+	protected void onCreateWithNonToolbarUpButton(Bundle savedInstanceState) {
+		this.withNonToolbarUpButton = true;
 
 		super.onCreate(savedInstanceState);
-		
-		if (withUpButton) {
-			ActionBar actionBar = getActionBar();
-			if (actionBar != null) {
-				actionBar.setDisplayHomeAsUpEnabled(true);
-			}
+
+		final ActionBar actionBar = getSupportActionBar();
+		if (actionBar != null) {
+			actionBar.setDisplayHomeAsUpEnabled(true);
 		}
 	}
-	
+
 	@Override public boolean onOptionsItemSelected(MenuItem item) {
-		if (withUpButton && item.getItemId() == android.R.id.home) {
-			finish();
-            return true;
+		if (withNonToolbarUpButton && item.getItemId() == android.R.id.home) {
+			navigateUp();
+			return true;
 		}
-		
+
 		return super.onOptionsItemSelected(item);
 	}
-	
-	@Override protected void onStart() {
-		super.onStart();
-	    EasyTracker.getInstance().activityStart(this);
-	}
-	
-	@Override protected void onStop() {
-		super.onStop();
-		EasyTracker.getInstance().activityStop(this);
+
+    protected void navigateUp() {
+		final Intent upIntent = NavUtils.getParentActivityIntent(this);
+		if (upIntent == null) { // not defined in manifest, let us finish() instead.
+			finish();
+			return;
+		}
+
+		if (NavUtils.shouldUpRecreateTask(this, upIntent)) {
+			TaskStackBuilder.create(this)
+				.addNextIntentWithParentStack(upIntent)
+				.startActivities();
+		} else {
+			NavUtils.navigateUpTo(this, upIntent);
+		}
 	}
 }
