@@ -6,10 +6,8 @@ import android.content.ContentResolver;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.v4.util.ArrayMap;
 import android.util.Log;
-import android.util.Pair;
 import com.google.gson.JsonIOException;
 import com.google.gson.JsonSyntaxException;
 import com.squareup.okhttp.Call;
@@ -18,16 +16,10 @@ import com.squareup.okhttp.Request;
 import com.squareup.okhttp.RequestBody;
 import yuku.afw.storage.Preferences;
 import yuku.alkitab.base.App;
-import yuku.alkitab.base.S;
-import yuku.alkitab.base.U;
 import yuku.alkitab.base.model.SyncShadow;
 import yuku.alkitab.base.storage.Prefkey;
-import yuku.alkitab.base.util.Sqlitil;
 import yuku.alkitab.debug.BuildConfig;
 import yuku.alkitab.debug.R;
-import yuku.alkitab.model.Label;
-import yuku.alkitab.model.Marker;
-import yuku.alkitab.model.Marker_Label;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -41,13 +33,8 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import static yuku.alkitab.base.util.Literals.List;
-
 public class Sync {
 	static final String TAG = Sync.class.getSimpleName();
-
-	/** From developer console: Client ID for web application */
-	public static final String CLIENT_ID = "979768394162-d3ev7tnali57es0tca97snvl24d2jbl2.apps.googleusercontent.com";
 
 	/**
 	 * The reason we use an installation id instead of just the simpleToken
@@ -67,11 +54,21 @@ public class Sync {
 		add, mod, del, // do not change the enum value names here. This will be un/serialized by gson.
 	}
 
+	public enum ApplyAppendDeltaResult {
+		ok,
+		unknown_kind,
+		/** Entities have changed during sync request */
+		dirty_entities,
+		/** Sync user account has changed during sync request */
+		dirty_sync_account,
+	}
+
 	public static class Operation<C> {
 		public Opkind opkind;
 		public String kind;
 		public String gid;
 		public C content;
+		public String creator_id; // for now, only used when receiving from server, not sending
 
 		public Operation(final Opkind opkind, final String kind, final String gid, final C content) {
 			this.opkind = opkind;
@@ -109,8 +106,11 @@ public class Sync {
 		public static final String KIND_MARKER = "Marker";
 		public static final String KIND_LABEL = "Label";
 		public static final String KIND_MARKER_LABEL = "Marker_Label";
+		public static final String KIND_HISTORY_ENTRY = "HistoryEntry";
 
-		/** Kind of this entity. Currently can be {@link #KIND_MARKER}, {@link #KIND_LABEL}, {@link #KIND_MARKER_LABEL}. */
+		/**
+		 * Kind of this entity. One of the <code>KIND_</code> constants on {@link yuku.alkitab.base.sync.Sync.Entity}.
+		 */
 		public String kind;
 		public String gid;
 		public C content;
@@ -138,277 +138,6 @@ public class Sync {
 			return result;
 		}
 		//endregion
-	}
-
-	/**
-	 * Entity content for {@link yuku.alkitab.model.Marker} and {@link yuku.alkitab.model.Label}.
-	 */
-	public static class MabelContent {
-		public Integer ari; // marker
-		public Integer kind; // marker
-		public String caption; // marker
-		public Integer verseCount; // marker
-		public Integer createTime; // marker
-		public Integer modifyTime; // marker
-		public String title; // label
-		public Integer ordering; // label
-		public String backgroundColor; // label
-		public String marker_gid; // marker_label
-		public String label_gid; // marker_label
-
-		//region boilerplate equals and hashCode methods
-		@Override
-		public boolean equals(final Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			final MabelContent that = (MabelContent) o;
-
-			if (ari != null ? !ari.equals(that.ari) : that.ari != null) return false;
-			if (backgroundColor != null ? !backgroundColor.equals(that.backgroundColor) : that.backgroundColor != null) return false;
-			if (caption != null ? !caption.equals(that.caption) : that.caption != null) return false;
-			if (createTime != null ? !createTime.equals(that.createTime) : that.createTime != null) return false;
-			if (kind != null ? !kind.equals(that.kind) : that.kind != null) return false;
-			if (label_gid != null ? !label_gid.equals(that.label_gid) : that.label_gid != null) return false;
-			if (marker_gid != null ? !marker_gid.equals(that.marker_gid) : that.marker_gid != null) return false;
-			if (modifyTime != null ? !modifyTime.equals(that.modifyTime) : that.modifyTime != null) return false;
-			if (ordering != null ? !ordering.equals(that.ordering) : that.ordering != null) return false;
-			if (title != null ? !title.equals(that.title) : that.title != null) return false;
-			if (verseCount != null ? !verseCount.equals(that.verseCount) : that.verseCount != null) return false;
-
-			return true;
-		}
-
-		@Override
-		public int hashCode() {
-			int result = ari != null ? ari.hashCode() : 0;
-			result = 31 * result + (kind != null ? kind.hashCode() : 0);
-			result = 31 * result + (caption != null ? caption.hashCode() : 0);
-			result = 31 * result + (verseCount != null ? verseCount.hashCode() : 0);
-			result = 31 * result + (createTime != null ? createTime.hashCode() : 0);
-			result = 31 * result + (modifyTime != null ? modifyTime.hashCode() : 0);
-			result = 31 * result + (title != null ? title.hashCode() : 0);
-			result = 31 * result + (ordering != null ? ordering.hashCode() : 0);
-			result = 31 * result + (backgroundColor != null ? backgroundColor.hashCode() : 0);
-			result = 31 * result + (marker_gid != null ? marker_gid.hashCode() : 0);
-			result = 31 * result + (label_gid != null ? label_gid.hashCode() : 0);
-			return result;
-		}
-		//endregion
-
-		@Override
-		public String toString() {
-			final StringBuilder sb = new StringBuilder("{");
-			if (ari != null) sb.append(ari).append(' ');
-			if (kind != null) sb.append(kind).append(' ');
-			if (caption != null) sb.append(q(caption)).append(' ');
-			if (verseCount != null) sb.append(verseCount).append(' ');
-			if (createTime != null) sb.append(createTime).append(' ');
-			if (modifyTime != null) sb.append(modifyTime).append(' ');
-			if (title != null) sb.append(q(title)).append(' ');
-			if (ordering != null) sb.append(ordering).append(' ');
-			if (backgroundColor != null) sb.append(backgroundColor).append(' ');
-			if (marker_gid != null) sb.append(marker_gid.substring(0, 10)).append(' ');
-			if (label_gid != null) sb.append(label_gid.substring(0, 10)).append(' ');
-
-			sb.setLength(sb.length() - 1);
-			sb.append('}');
-			return sb.toString();
-		}
-
-		@NonNull static String q(@NonNull String s) {
-			final String c;
-			if (s.length() > 20) {
-				c = s.substring(0, 19).replace("\n", "\\n") + "…";
-			} else {
-				c = s.replace("\n", "\\n");
-			}
-			return "'" + c + "'";
-		}
-	}
-
-	public static class SyncShadowMabelDataJson {
-		public List<Entity<MabelContent>> entities;
-	}
-
-	public static class MabelClientState {
-		public final int base_revno;
-		@NonNull public final Delta<MabelContent> delta;
-
-		public MabelClientState(final int base_revno, @NonNull final Delta<MabelContent> delta) {
-			this.base_revno = base_revno;
-			this.delta = delta;
-		}
-	}
-
-	/**
-	 * @return base revno, delta of shadow -> current.
-	 */
-	public static Pair<MabelClientState, List<Entity<MabelContent>>> getMabelClientStateAndCurrentEntities() {
-		final SyncShadow ss = S.getDb().getSyncShadowBySyncSetName(SyncShadow.SYNC_SET_MABEL);
-
-		final List<Entity<MabelContent>> srcs = ss == null? List(): mabelEntitiesFromShadow(ss);
-		final List<Entity<MabelContent>> dsts = getMabelEntitiesFromCurrent();
-
-		final Delta<MabelContent> delta = new Delta<>();
-
-		// additions and modifications
-		for (final Entity<MabelContent> dst : dsts) {
-			final Entity<MabelContent> existing = findMabelEntity(srcs, dst.gid, dst.kind);
-
-			if (existing == null) {
-				delta.operations.add(new Operation<>(Opkind.add, dst.kind, dst.gid, dst.content));
-			} else {
-				if (!isSameMabelContent(dst, existing)) { // only when it changes
-					delta.operations.add(new Operation<>(Opkind.mod, dst.kind, dst.gid, dst.content));
-				}
-			}
-		}
-
-		// deletions
-		for (final Entity<MabelContent> src : srcs) {
-			final Entity<MabelContent> still_have = findMabelEntity(dsts, src.gid, src.kind);
-			if (still_have == null) {
-				delta.operations.add(new Operation<>(Opkind.del, src.kind, src.gid, null));
-			}
-		}
-
-		return Pair.create(new MabelClientState(ss == null ? 0 : ss.revno, delta), dsts);
-	}
-
-	private static boolean isSameMabelContent(final Entity<MabelContent> a, final Entity<MabelContent> b) {
-		if (!U.equals(a.gid, b.gid)) return false;
-		if (!U.equals(a.kind, b.kind)) return false;
-
-		return U.equals(a.content, b.content);
-	}
-
-	private static Entity<MabelContent> findMabelEntity(final List<Entity<MabelContent>> list, final String gid, final String kind) {
-		for (final Entity<MabelContent> entity : list) {
-			if (U.equals(gid, entity.gid) && U.equals(kind, entity.kind)) {
-				return entity;
-			}
-		}
-		return null;
-	}
-
-	private static List<Entity<MabelContent>> mabelEntitiesFromShadow(@NonNull final SyncShadow ss) {
-		final SyncShadowMabelDataJson data = App.getDefaultGson().fromJson(U.utf8BytesToString(ss.data), SyncShadowMabelDataJson.class);
-		return data.entities;
-	}
-
-	@NonNull public static SyncShadow shadowFromMabelEntities(@NonNull final List<Entity<MabelContent>> entities, final int revno) {
-		final SyncShadowMabelDataJson data = new SyncShadowMabelDataJson();
-		data.entities = entities;
-		final String s = App.getDefaultGson().toJson(data);
-		final SyncShadow res = new SyncShadow();
-		res.data = U.stringToUtf8Bytes(s);
-		res.syncSetName = SyncShadow.SYNC_SET_MABEL;
-		res.revno = revno;
-		return res;
-	}
-
-	@NonNull public static List<Entity<MabelContent>> getMabelEntitiesFromCurrent() {
-		final List<Entity<MabelContent>> res = new ArrayList<>();
-
-		{ // markers
-			for (final Marker marker : S.getDb().listAllMarkers()) {
-				final Entity<MabelContent> entity = new Entity<>();
-				entity.kind = Entity.KIND_MARKER;
-				entity.gid = marker.gid;
-				final MabelContent content = entity.content = new MabelContent();
-				content.ari = marker.ari;
-				content.caption = marker.caption;
-				content.kind = marker.kind.code;
-				content.verseCount = marker.verseCount;
-				content.createTime = Sqlitil.toInt(marker.createTime);
-				content.modifyTime = Sqlitil.toInt(marker.modifyTime);
-				res.add(entity);
-			}
-		}
-
-		{ // labels
-			for (final Label label : S.getDb().listAllLabels()) {
-				final Entity<MabelContent> entity = new Entity<>();
-				entity.kind = Entity.KIND_LABEL;
-				entity.gid = label.gid;
-				final MabelContent content = entity.content = new MabelContent();
-				content.title = label.title;
-				content.backgroundColor = label.backgroundColor;
-				content.ordering = label.ordering;
-				res.add(entity);
-			}
-		}
-
-		{ // marker_labels
-			for (final Marker_Label marker_label : S.getDb().listAllMarker_Labels()) {
-				final Entity<MabelContent> entity = new Entity<>();
-				entity.kind = Entity.KIND_MARKER_LABEL;
-				entity.gid = marker_label.gid;
-				final MabelContent content = entity.content = new MabelContent();
-				content.marker_gid = marker_label.marker_gid;
-				content.label_gid = marker_label.label_gid;
-				res.add(entity);
-			}
-		}
-
-		return res;
-	}
-
-	/**
-	 * Modify or create a marker from an entity content. This is called when the server append delta
-	 * asks for an add or a mod operation.
-	 * This will not merge content, will only overwrite.
-	 * @param marker an existing marker (content will be modified), or null to create a new marker
-	 * @param content entity content, containing the new data.
-	 */
-	@NonNull public static Marker updateMarkerWithEntityContent(@Nullable final Marker marker, @NonNull final String gid, @NonNull final MabelContent content) {
-		final Marker res = marker != null ? marker : Marker.createEmptyMarker();
-
-		res.gid = gid;
-		res.ari = content.ari;
-		res.kind = Marker.Kind.fromCode(content.kind);
-		res.caption = content.caption;
-		res.verseCount = content.verseCount;
-		res.createTime = Sqlitil.toDate(content.createTime);
-		res.modifyTime = Sqlitil.toDate(content.modifyTime);
-
-		return res;
-	}
-
-	/**
-	 * Modify or create a label from an entity content. This is called when the server append delta
-	 * asks for an add or a mod operation.
-	 * This will not merge content, will only overwrite.
-	 * @param label an existing label (content will be modified), or null to create a new label
-	 * @param content entity content, containing the new data.
-	 */
-	@NonNull public static Label updateLabelWithEntityContent(@Nullable final Label label, @NonNull final String gid, @NonNull final MabelContent content) {
-		final Label res = label != null ? label : Label.createEmptyLabel();
-
-		res.gid = gid;
-		res.title = content.title;
-		res.ordering = content.ordering;
-		res.backgroundColor = content.backgroundColor;
-
-		return res;
-	}
-
-	/**
-	 * Modify or create a marker-label association from an entity content. This is called when the server append delta
-	 * asks for an add or a mod operation.
-	 * This will not merge content, will only overwrite.
-	 * @param marker_label an existing marker-label association (content will be modified), or null to create a new marker-label association
-	 * @param content entity content, containing the new data.
-	 */
-	@NonNull public static Marker_Label updateMarker_LabelWithEntityContent(@Nullable final Marker_Label marker_label, @NonNull final String gid, @NonNull final MabelContent content) {
-		final Marker_Label res = marker_label != null ? marker_label : Marker_Label.createEmptyMarker_Label();
-
-		res.gid = gid;
-		res.marker_gid = content.marker_gid;
-		res.label_gid = content.label_gid;
-
-		return res;
 	}
 
 	/**
@@ -545,6 +274,7 @@ public class Sync {
 	public static boolean sendGcmRegistrationId(final String simpleToken, final String registration_id) {
 		final RequestBody requestBody = new FormEncodingBuilder()
 			.add("simpleToken", simpleToken)
+			.add("sender_id", Gcm.SENDER_ID)  // not really needed, but for logging on server
 			.add("registration_id", registration_id)
 			.build();
 
@@ -789,7 +519,7 @@ public class Sync {
 		}
 
 		// request sync.
-		for (final String syncSetName : SyncShadow.ALL_SYNC_SETS) {
+		for (final String syncSetName : SyncShadow.ALL_SYNC_SET_NAMES) {
 			final Bundle extras = new Bundle();
 			extras.putString(SyncAdapter.EXTRA_SYNC_SET_NAME, syncSetName);
 			extras.putBoolean(ContentResolver.SYNC_EXTRAS_MANUAL, true);
